// src/components/BCol/BCol.interface.ts
interface BColInterface {
  alignSelf: string
  col: boolean
  cols?: string | number
  lg?: boolean | number | string
  md?: boolean | number | string
  sm?: boolean | number | string
  xl?: boolean | number | string
  offset?: number | string
  offsetLg?: number | string
  offsetMd?: number | string
  offsetSm?: number | string
  offsetXl?: number | string
  order?: number | string
  orderLg?: number | string
  orderMd?: number | string
  orderSm?: number | string
  orderXl?: number | string
}

export default BColInterface